% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{Greeting}
\usage{
helloworld(name = "World")
}
\arguments{
\item{name}{String - defaults to World.}
}
\value{
nothing
}
\description{
Prints 'Hello \if{html}{\out{<name>}}!'.
}
\examples{
helloworld("Jacob")
}
